refactor: 가격 타입 int로 변경
- 시간 제한 안에 풀어야하므로 우선 구현시에 int 타입으로 정하기
- Price 값 객체는 유효성만 검사하고 계산시에는 값으로 계산하기

+++ 4-period/java-vendingmachine-precourse-main/src/main/java/vendingmachine/controller/VendingController.java
@@ -33,11 +33,11 @@ public class VendingController {
 
     public void process() {
         Price holdingPrice = createHoldingPrice();
-        Map<Coin, Long> coins = service.createRandomCoins(holdingPrice);
+        Map<Coin, Integer> coins = service.createRandomCoins(holdingPrice);
         List<CoinDto> coinDtos = new ArrayList<>();
         for (Coin coin : Coin.values()) {
             if (coins.containsKey(coin)) {
-                Long quantity = coins.get(coin);
+                int quantity = coins.get(coin);
                 coinDtos.add(new CoinDto(coin.getPrice().getAmount(), quantity));
                 continue;
             }
@@ -58,7 +58,7 @@ public class VendingController {
         }
         Price remainingPrice = orderProcessor.getInputPrice();
         LeastCoinGenerator leastCoinGenerator = new LeastCoinGenerator(coins);
-        Map<Coin, Long> leastCoins = leastCoinGenerator.generateCoins(remainingPrice);
+        Map<Coin, Integer> leastCoins = leastCoinGenerator.generateCoins(remainingPrice);
         outputView.showRemainingPrice(leastCoins);
     }
 
@@ -70,7 +70,7 @@ public class VendingController {
     private Price createInputPrice() {
         outputView.requestInputAmount();
         return exceptionHandler.retryOn(() -> {
-            long inputPrice = inputView.readInputPrice();
+            int inputPrice = inputView.readInputPrice();
             return service.createInputPrice(inputPrice);
         });
     }
@@ -86,7 +86,7 @@ public class VendingController {
     private Price createHoldingPrice() {
         outputView.requestHoldingAmount();
         return exceptionHandler.retryOn(() -> {
-            long holdingAmount = inputView.readHoldingAmount();
+            int holdingAmount = inputView.readHoldingAmount();
             return service.createHoldingPrice(holdingAmount);
         });
     }
